Crate miden_mast_package

The Package containing a Program or Library and a manifest consisting of its exports and dependencies.

Modules

debug_info

Debug information sections for MASP packages.

Structs

ConstantExport

A constant definition exported by a package

Dependency

A package dependency

DependencyName

The name of a dependency

InvalidPackageKindError

Error returned when trying to convert an invalid u8 to a PackageKind.

Library

Represents a library where all modules were compiled into a MastForest.

MastForest

Represents one or more procedures, represented as a collection of [MastNode]s.

MemDependencyResolverByDigest

Package

A package containing a Program/Library, and a manifest (exports and dependencies).

PackageManifest

The manifest of a package, containing the set of package dependencies (libraries or packages) and exported items (procedures, constants, types), if known.

PathBuf

Path to an item in a library, i.e. module, procedure, constant or type.

ProcedureExport

A procedure exported by a package, along with its digest, signature, and attributes.

ProcedureName

Procedure name.

Program

An executable program for Miden VM.

QualifiedProcedureName

Represents a qualified procedure name, e.g. std::math::u64::add, parsed into it’s constituent [Path] and ProcedureName components.

Section

SectionId

A unique identifier for optional sections of the Miden package format

TypeExport

A named type declaration exported by a package

Version

SemVer version as defined by https://semver.org.

VersionError

Error parsing a SemVer version or version requirement.

Word

A unit of data consisting of 4 field elements.

Enums

InvalidSectionIdError

LocalResolvedDependency

Resolved local(loaded) dependency

MastArtifact

The artifact produced by lowering a program or library to a Merkelized Abstract Syntax Tree

PackageExport

Represents a named item exported from a package.

PackageKind

The kind of project that produced this package.

ResolvedDependency

Resolved dependency In the future, we intend to add Registry variant to describe dependencies to be loaded from the registry.

Traits

DependencyResolver